refactor: NetworkManagerError 제거, DataDecoder에 디코더 병합
